fix: carry the frame-rate target as an int on both platforms
The channel value became an int with StreamFrameRate but the two native
consumers kept their Double fields from the old API, leaving toDouble()
and Double() bridges at the call sites.

+++ android/src/main/kotlin/io/rodcone/flutter_meta_wearables_dat/MetaWearablesDatPlugin.kt
@@ -1249,7 +1249,7 @@ class MetaWearablesDatPlugin :
                 }
 
                 sessionKey = key
-                frameProcessor.configure(fps.toDouble())
+                frameProcessor.configure(fps)
 
                 // Register a Flutter texture for zero-copy rendering
                 val registry = textureRegistry
